Move the db file to utils folder
